# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * MIME type of the image
 */
export type ImageInputMimeType = typeof ImageInputMimeType[keyof typeof ImageInputMimeType];


export const ImageInputMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'image/gif': 'image/gif',
} as const;

export interface ImageInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type of the image */
  mimeType: ImageInputMimeType;
}

export interface ExtractedText {
  /** The text extracted from the image */
  text: string;
}

export interface PassageInput {
  /**
     * The SAT reading passage text to visualize
     * @minLength 50
     * @maxLength 3000
     */
  passage: string;
}

export interface Scene {
  /** Scene number (0-based) */
  index: number;
  /** Short scene title */
  title: string;
  /** Visual description for image generation */
  description: string;
  /** Caption text to overlay on the video */
  caption: string;
}

export type VisualizeResultStatus = typeof VisualizeResultStatus[keyof typeof VisualizeResultStatus];


export const VisualizeResultStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface VisualizeResult {
  /** Unique job identifier for polling */
  jobId: string;
  status: VisualizeResultStatus;
  scenes: Scene[];
}

export type JobStatusStatus = typeof JobStatusStatus[keyof typeof JobStatusStatus];


export const JobStatusStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  /** Completion percentage 0-100 */
  progress: number;
  /**
     * Current processing step label
     * @nullable
     */
  step?: string | null;
  scenes?: Scene[];
  /** Base64-encoded JPEG thumbnail for each scene as it is generated (one per completed scene) */
  thumbnails?: string[];
  /**
     * Error message if status is error
     * @nullable
     */
  error?: string | null;
}

export interface ErrorResult {
  error: string;
}