# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusStatus } from './jobStatusStatus';
import type { Scene } from './scene';

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  /** Completion percentage 0-100 */
  progress: number;
  /**
     * Current processing step label
     * @nullable
     */
  step?: string | null;
  scenes?: Scene[];
  /** Base64-encoded JPEG thumbnail for each scene as it is generated (one per completed scene) */
  thumbnails?: string[];
  /**
     * Error message if status is error
     * @nullable
     */
  error?: string | null;
}